Modify counter data struct to eliminate false sharing issue with cache
So, cache lines are fetched in aligned 64 byte chunks, and it pulls
any other variable with it. When this program had an array of int64s,
there were many threads simultaneously contending over a small number
of cache lines, and it was playing hell with cache coherency.

Fun fact: Intel handles false sharing WAY better than AMD does, and
on intel machines, there was a slowdown by a factor of 4x-ish, and on
Zen machines, the slowdown is closer to a factor of 20x+